+++ __snapshots__/bootstrapper.js
@@ -4,6 +4,7 @@ exports['Bootstrapper should open a PR 1'] = `
     ".": {
       "release-type": "node"
     }
-  }
+  },
+  "$schema": "https://raw.githubusercontent.com/googleapis/release-please/main/schemas/config.json"
 }
 `